lib: move DeploymentID tests to own file
