+++ lib/deployment_test.go
@@ -1,7 +1,6 @@
 package sous
 
 import (
-	"fmt"
 	"reflect"
 	"testing"
 
@@ -131,26 +130,6 @@ func TestBuildDeployment(t *testing.T) {
 			assert.Equal("c", d.DeployConfig.Volumes[0].Container)
 		}
 		assert.Equal(nick, d.ClusterName)
-	}
-}
-
-func TestDigest(t *testing.T) {
-	tmpl := "got:%s expected:%s"
-	expected := "3ea161adca77a01781628e8a7d24ad0e"
-	d := &DeploymentID{
-		ManifestID: ManifestID{
-			Source: SourceLocation{
-				Repo: "fake.tld/org/" + "project",
-				Dir:  "down/here",
-			},
-		},
-		Cluster: "test-cluster",
-	}
-	dStr := fmt.Sprintf("%x", d.Digest())
-	if dStr != expected {
-		t.Fatalf(tmpl, dStr, expected)
-	} else {
-		t.Logf(tmpl, dStr, expected)
 	}
 }
